Add validation for BazarAdmin permissions and update related models

+++ src/apps/bazars/services/delete_place.py
@@ -1,12 +1,15 @@
 from rest_framework.exceptions import NotFound
 from apps.bazars.models import Place
+from apps.core.services.validate_bazar_admin import validate_bazar_admin
 
 
-def delete_place(place_id: int):
+def delete_place(user, place_id: int):
     try:
         place = Place.objects.get(id=place_id)
     except Place.DoesNotExist:
         raise NotFound({"message_key": "place_not_found", "message": "Place not found"})
+
+    validate_bazar_admin(user, place.bazar_id)
 
     bazar = place.bazar
     place.delete()
